Agregue un poco de navegabilidad entre las pantallas y agregue un ventana emergente para las acciones de borrar.

+++ AgendaPlusUWP/Views/FechasImportantes/FechasImportantesMain.xaml.cs
@@ -31,9 +31,17 @@ namespace AgendaPlusUWP.Views.FechasImportantes
 
         public FechasImportantesMain()
         {
-            userID = 1;
             this.InitializeComponent();
             inizializarAPI();
+        }
+
+        protected override void OnNavigatedTo(NavigationEventArgs e)
+        {
+            string IDstr = e.Parameter.ToString();
+
+            userID = Int32.Parse(IDstr);
+
+            base.OnNavigatedTo(e);
         }
 
         private async void inizializarAPI()
@@ -81,13 +89,34 @@ namespace AgendaPlusUWP.Views.FechasImportantes
             {
                 fecha.Usuario = null;
 
-                var httpHandler = new HttpClientHandler();
-                var client = new HttpClient(httpHandler);
-                var json = JsonConvert.SerializeObject(fecha);
-                var content = new StringContent(json, System.Text.Encoding.UTF8, "application/json");
-                HttpResponseMessage response = await client.DeleteAsync($"https://localhost:44304/api/fechasimportantes/{fecha.FechasImportantesID}");
-                inizializarAPI();
+                ContentDialog noWifiDialog = new ContentDialog
+                {
+                    Title = "Advertencia",
+                    Content = "estas seguro de que quieres borrar la fecha",
+                    CloseButtonText = "Cancelar",
+                    PrimaryButtonText = "Borrar"
+                };
+
+                ContentDialogResult result = await noWifiDialog.ShowAsync();
+                string resultSTR = result.ToString();
+
+                if (resultSTR.Equals("Primary"))
+                {
+
+                    var httpHandler = new HttpClientHandler();
+                    var client = new HttpClient(httpHandler);
+                    var json = JsonConvert.SerializeObject(fecha);
+                    var content = new StringContent(json, System.Text.Encoding.UTF8, "application/json");
+                    HttpResponseMessage response = await client.DeleteAsync($"https://localhost:44304/api/fechasimportantes/{fecha.FechasImportantesID}");
+                    inizializarAPI();
+                }
             }
+        }
+
+        private void crearFecha(object sender, RoutedEventArgs e)
+        {
+            Frame.Content = null;
+            Frame.Navigate(typeof(FechasImportantesCreate), userID);
         }
     }
 }
